+++ lib/kitchen/directions/bake_chapter_glossary/v1.rb
@@ -24,10 +24,6 @@ module Kitchen::Directions::BakeChapterGlossary
     end
 
     def bake(chapter:, metadata_source:, append_to: nil, uuid_prefix: '')
-      @metadata = metadata_source.children_to_keep.copy
-      @klass = 'glossary'
-      @title = I18n.t(:eoc_key_terms_title)
-      @uuid_prefix = uuid_prefix
       @glossary = []
 
       chapter.glossaries.search('dl').each do |definition_element|
@@ -36,13 +32,15 @@ module Kitchen::Directions::BakeChapterGlossary
 
       chapter.glossaries.trash
 
-      @content = @glossary.sort.map { |definition| definition.element.paste }.join
+      content = @glossary.sort.map { |definition| definition.element.paste }.join
 
-      append_to_element = append_to || chapter
-      @in_composite_chapter = append_to_element.is?(:composite_chapter)
-
-      append_to_element.append(child: render(file:
-        '../../templates/eoc_section_title_template.xhtml.erb'))
+      Kitchen::Directions::EocCompositePageContainer.v1(
+        container_key: 'glossary',
+        uuid_key: "#{uuid_prefix}glossary",
+        metadata_source: metadata_source,
+        content: content,
+        append_to: append_to || chapter
+      )
     end
   end
 end