+++ lib/kitchen/directions/bake_chapter_references/v1.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 module Kitchen::Directions::BakeChapterReferences
+  # TODO: Refactor to use EocCompositePageContainer or MoveCustomSectionToEocContainer
   class V1
     renderable
 
@@ -20,7 +21,7 @@ module Kitchen::Directions::BakeChapterReferences
 
       @content = chapter.pages.references.cut.paste
       chapter.append(child: render(file:
-        '../../templates/eoc_section_title_template.xhtml.erb'))
+        '../../templates/eoc_section_template_old.xhtml.erb'))
     end
 
     def bake_page_references(page:)