+++ lib/kitchen/directions/bake_chapter_summary.rb
@@ -17,50 +17,29 @@ module Kitchen
       class V1
         renderable
         def bake(chapter:, metadata_source:, uuid_prefix: '.', klass: 'summary')
-          @metadata = metadata_source.children_to_keep.copy
-          @klass = klass
-          @title = I18n.t(:eoc_summary_title)
-          @uuid_prefix = uuid_prefix
-
           summaries = Clipboard.new
 
-          # TODO: include specific page types somehow without writing it out
           chapter.non_introduction_pages.each do |page|
             summary = page.summary
 
             next if summary.nil?
 
             summary.first("[data-type='title']")&.trash # get rid of old title if exists
-            summary_title = page.title.copy
-            summary_title.name = 'h3'
-
-            unless summary_title.children.search('span.os-number').present?
-              summary_title.replace_children(with:
-                <<~HTML
-                  <span class="os-number">#{chapter.count_in(:book)}.#{page.count_in(:chapter)}</span>
-                  <span class="os-divider"> </span>
-                  <span class="os-text" data-type="" itemprop="">#{summary_title.children}</span>
-                HTML
-              )
-            end
-
-            summary.prepend(child:
-              <<~HTML
-                <a href="##{page.title.id}">
-                  #{summary_title.paste}
-                </a>
-              HTML
-            )
+            title = EocSectionTitleLinkSnippet.v1(page: page)
+            summary.prepend(child: title)
+            summary.first('h3')[:itemprop] = 'name'
             summary.cut(to: summaries)
           end
 
           return if summaries.none?
 
-          @content = summaries.paste
-          @in_composite_chapter = false
-
-          chapter.append(child: render(file:
-            '../templates/eoc_section_title_template.xhtml.erb'))
+          EocCompositePageContainer.v1(
+            container_key: klass,
+            uuid_key: "#{uuid_prefix}#{klass}",
+            metadata_source: metadata_source,
+            content: summaries.paste,
+            append_to: chapter
+          )
         end
       end
     end