+++ lib/kitchen/directions/bake_eoc_section_content/remove_section_title.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Kitchen
+  module Directions
+    module RemoveSectionTitle
+      def self.v1(section:)
+        section.first('[data-type="title"]')&.trash
+      end
+    end
+  end
+end