+++ lib/kitchen/directions/bake_further_research.rb
@@ -15,45 +15,20 @@ module Kitchen
       class V1
         renderable
         def bake(chapter:, metadata_source:, uuid_prefix: '.')
-          @metadata = metadata_source.children_to_keep.copy
-          @klass = 'further-research'
-          @title = I18n.t(:eoc_further_research_title)
-          @uuid_prefix = uuid_prefix
-
-          further_researches = Clipboard.new
-
-          chapter.non_introduction_pages.each do |page|
-            further_research = page.first('.further-research')
-            next unless further_research.present?
-
-            further_research.first("[data-type='title']")&.trash # get rid of old title if exists
-            further_research_title = page.title.copy
-            further_research_title.name = 'h3'
-            further_research_title.replace_children(with: <<~HTML
-              <span class="os-number">#{chapter.count_in(:book)}.#{page.count_in(:chapter)}</span>
-              <span class="os-divider"> </span>
-              <span class="os-text" data-type="" itemprop="">#{further_research_title.children}</span>
-            HTML
-            )
-
-            further_research.prepend(child:
-              <<~HTML
-                <a href="##{page.title.id}">
-                  #{further_research_title.paste}
-                </a>
-              HTML
-            )
-            further_research.cut(to: further_researches)
+          MoveCustomSectionToEocContainer.v1(
+            chapter: chapter,
+            metadata_source: metadata_source,
+            container_key: 'further-research',
+            uuid_key: "#{uuid_prefix}further-research",
+            section_selector: 'section.further-research',
+            append_to: nil,
+            include_intro_page: false
+          ) do |further_research|
+            RemoveSectionTitle.v1(section: further_research)
+            title = EocSectionTitleLinkSnippet.v1(page: further_research.ancestor(:page))
+            further_research.prepend(child: title)
+            further_research.first('h3')[:itemprop] = 'name'
           end
-
-          return if further_researches.none?
-
-          @content = further_researches.paste
-
-          @in_composite_chapter = false
-
-          chapter.append(child: render(file:
-            '../templates/eoc_section_title_template.xhtml.erb'))
         end
       end
     end