+++ spec/directions/move_exercises_to_eoc/v1_spec.rb
@@ -69,7 +69,7 @@ RSpec.describe Kitchen::Directions::MoveExercisesToEOC::V1 do
                     <span class="os-text">foo</span>
                   </h3>
                   <div data-type="metadata" style="display: none;">
-                    <h1 data-type="document-title" itemprop="name">Exercises</h1>
+                    <h1 data-type="document-title" itemprop="name">foo</h1>
                     <div class="authors" id="authors_copy_1">Authors</div>
                     <div class="publishers" id="publishers_copy_1">Publishers</div>
                     <div class="print-style" id="print-style_copy_1">Print Style</div>