+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Add documentation to `MoveCustomSectionToEocContainer` & `EocCompositePageContainer` (minor)
+* Add a `MoveCustomSectionToEocContainer` to allow for custom sections (minor)
+* Add a `EocCompositePageContainer` direction to handle creation of eoc page wrapper (minor)
+* Refactor `BakeChapterKeyEquations`, `BakeChapterGlossary`, `BakeChapterSummary`, `BakeFurtherResearch` (major)
+* Fix metadata title in composite pages (major)
 
 ## [8.0.1] - 2021-06-29